[REFACTOR] Data 레이어를 data/attendance 모듈로 이동
- AttendanceService, MockAttendanceService 이동
- AttendanceRepositoryImpl 이동
- Response/Request 모델 클래스들 이동
- DI 모듈을 AttendanceDataModule로 리팩토링
