+++ data/attendance/src/main/java/org/sopt/official/data/attendance/service/MockAttendanceService.kt
@@ -22,10 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.data.service.attendance
+package org.sopt.official.data.attendance.service
 
 import kotlinx.serialization.json.Json
-import org.sopt.official.data.model.attendance.*
+import org.sopt.official.data.attendance.model.*
 
 class MockAttendanceService : AttendanceService {
     override suspend fun getSoptEvent(): BaseAttendanceResponse<SoptEventResponse> {
